ast generator fmu work in progress
