+++ numerous/engine/system/fmu_system_generator/utils.py
@@ -1,0 +1,13 @@
+from numba.core.extending import intrinsic
+
+
+@intrinsic
+def address_as_void_pointer(typingctx, src):
+    """ returns a void pointer from a given memory address """
+    from numba.core import types, cgutils
+    sig = types.voidptr(src)
+
+    def codegen(cgctx, builder, sig, args):
+        return builder.inttoptr(args[0], cgutils.voidptr_t)
+
+    return sig, codegen
